Use size_t and std::vector for buffers
Replace signed int indices in http::getPath with size_t to fix signed/unsigned mismatches and potential negative index issues. In Networking::ClientSocket::recv, include <vector>, remove the DEFAULT_BUFLEN macro and raw C array, and use a std::vector<char> buffer with buffer.data() for safer, RAII-managed receive storage. Returns behavior preserved.

+++ src/http/http.cpp
@@ -4,10 +4,10 @@ namespace http
 {
     std::string getPath(std::string message)
     {
-        int start = message.find(' ');
-        int end = message.find(' ', start+1);
+        size_t start = message.find(' ');
+        size_t end = message.find(' ', start+1);
         std::string path;
-        for(int i = start+1; i < end; ++i)
+        for(size_t i = start+1; i < end; ++i)
             path += message[i];
         return path;
     }
